==> 01-introduccion_python.ipynb <==
### Introducción

En este notebook (script), exploraremos algunos de los conceptos fundamentales de Python para familiarizarnos con este lenguaje.


En cada notebook dedicaremos la celda inicial a la carga de librerías necesarias para la ejecución del script completo. Si necesitamos instalar algunas de ellas porque no estén disponibles en nuestro repositorio, las instalaremos y dejaremos indicado que hemos necesitado instalarlas. Esto servirá para la configuración del entorno de trabajo de futuros proyectos. :)


### Instalación de una nueva libreria (`numpy`):

pip install numpy

### Carga de librerias utiles (`numpy` y `pandas`):

import numpy as np
import pandas as pd

### Función `print()`

print('Bienvenid@s al seminario!')

###  Estructuras de datos
En Python existen múltiples estructuras en las que almacenar objetos:
- **List:** una lista mutable de elementos.
- **Tuple:** una estructura de datos de solo lectura (inmutable).
- **Dictionary:** utiliza una estructura de pares clave-valor para organizar y acceder a datos.

# Lista
programas = ['Python', "R", "Java", "Julia" ]

# Acceder al primer elemento de la lista
print(programas[0])

# Acceder al último elemento de la lista
print(programas[-1])

# Cambiar el valor del tercer elemento de la lista
programas[2]='C'
print(programas)

# Tupla
tupla = ('Python', "R", "Java", "Julia" )
print(tupla)

# Intentar cambiar el valor del tercer elemento de la tupla
tupla[2]='C'
print(tupla)

# Diccionario
dic = {'a': 1, 'b': 2, 'c': 3}

# Acceder al valor de la clave "b"
print(dic['b'])

### Pandas DataFrame

**Pandas**: Es una de las librerías más populares de Python. Proporciona las herramientas y estructuras necesarias para manipular y analizar datos. La estructura básica de Pandas es el **DataFrame**, una colección ordenada de columnas con nombres y tipos, donde una sola fila representa un único caso (observación) y las columnas representan atributos particulares.

- Guía: https://pandas.pydata.org/docs/getting_started/index.html

# Crear un DataFrame desde un diccionario
data = {'Nombre': ['Alice', 'Bob', 'Charlie', "Fredo", "Victor", "Krzyzs", "Raul"],
        'Edad': [25, 30, 35, 28, 33, 38, 23],
        'Ciudad': ['A', 'B', 'C', 'O', 'O', "M", "L"]}

df = pd.DataFrame(data)

# Ver el DataFrame
print(df)

# Comprobar sus dimensiones
print('Dimensiones:', df.shape)

# Imprimir las primeras 5 filas
df.head()

# Imprimir las ultimas 5 filas
df.tail()

# Seleccionar la columna Edad
df['Edad']

# Seleccionar el primer elemento
df.iloc[0]

# Imprimir el nombre del 5º elemento
print(df.loc[4, "Nombre"])
